Override PORTSDIR to the target ports tree on merge
PORTSDIR was previously not set, leading to different behavior depending
on the end user configuration (e.g. in /etc/make.conf).  Because
portshaker sometimes rely on make(1) to gather informations on a port,
it may fail to find the appropriate Mk/* files and report an error.

This fixes this situation.